datetime: use an int64_t instead of an int in timeval_addmsec
Since it deals with milliseconds, this function needs to handle large
numbers, for example one year does not fit in a simple int. Just fix it by
using an int64_t.

+++ datetime.h
@@ -366,7 +366,7 @@ static inline struct timeval timeval_add(struct timeval a, struct timeval b)
     return res;
 }
 
-static inline struct timeval timeval_addmsec(struct timeval a, int msec)
+static inline struct timeval timeval_addmsec(struct timeval a, int64_t msec)
 {
     struct timeval res;
